Attachment maximum file size
Add parameter to allow user to limit the attachment maximum file size
(in BYTES). If it is greater than the allowed max size, the attachment
won't be downloaded, it saves the bandwidth.
If attachment max size is blank, all attachment will be downloaded
without any restriction.
If attachment max size is less than or equal 0, no attachment will be
downloaded.